Added the beginnings of a cookbook.
The cookbook should be extended on an ongoing basic with relevant tasks
as suggested by biologists. Each new task should be included on the main
listing page.

To make it easy to check the validity of these code snippets, a script
has been added to the repository that reads a given file and runs each
code snippet. This should be run on each how-to before committing.